==> 00-preprocess-TCR.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# Parsing TCR data downloaded from Adaptive Biotech

Download the data as follows:
- navigate to <https://clients.adaptivebiotech.com/pub/dewolf-2022-gvhd>
- click explore this project
- log in (making an account if needed)
- click "Open in Analyses"
- click "Analyses on the nav pane
- select all
- click export -> export Sample (NOT  V2!)
- wait patiently
- continue waiting; resist the urge to try the process again
- eventually move downloaded zip to this repo, unpack, and move the tsv files to `data/raw_TCR_data/`.


```{R 00-preprocess-TCR-1 }
tcr_data_dir <- "data/raw_TCR_data/"
```

The resulting file `TCR/tcr_unified.csv` is used for later analyses including the GLIPH2 analysis.

```{r 00-preprocess-TCR-2 }
library(tidyverse)
knitr::opts_chunk$set(echo = TRUE)

dir.create("data/TCR", showWarnings = FALSE)


tcr_raw_output <- file.path("data/TCR", "tcr_unified.csv")
if (!file.exists(tcr_raw_output)){
  # TODO: ensure code works with dryad data pull. Until then, make the file as described in data/raw_TCR_data/README.md
  rdryad::dryad_download(dois = "10.5061/dryad.r4xgxd2k5")
  
}

```



## External Correlation Calculations

This can be done in R but the very large matrices make it slow. 
```
bsub -W 2:00 -n 128 -Is  bash
conda create -name tcr scipy pandas -y 
conda activate tcrs
bash scripts/get_file_pairs.sh ./data/raw_TCR_data/ > human_mouse_file_pairs.tsv
parallel -a human_mouse_file_pairs.tsv --bar --progress --colsep "\t" ~/miniconda3/envs/pd/bin/python scripts/AB-correlations.py {1} {2} > data/2022-08-29-tcr-correlations.tsv

```
### Parsing and processing the TCR Data

The following chunk serves two purposes: reading in the TCR data, and calculating the JSD correlations based on the TCR profiles.  This takes a while, so the results are saved in an object called `tcr_objects.Rdata`. If this object is not found, it will create it. Don't take the md5 warnings too seriously:  data downloaded on different machines can be reordered, which kinda breaks the point of this validation step.


```{R 00-preprocess-TCR-3}
tcr_objects_path <- "tcr_objects.Rdata"
if (!file.exists(tcr_objects_path)) {
  tcr_raw_output <- file.path("data/TCR", "tcr_unified.csv")
  if (tools::md5sum(tcr_raw_output) != "fd463374450fdbfca4565669dd49d89d") warning("tcr_unified.csv may need be re-created, it does not match the validated version")
  tcr_raw <- data.table::fread(tcr_raw_output) %>%
    filter(frame_type == "In") %>%
    as_tibble()

  tcr_sample_metadata <- tcr_raw %>%
    select(sample_name, org, samp_type, collection_day, sample_amount_ng, indv_type, exp_rep, gvhd, indv_id, indv_label, tissue, indv_tissue_label, species:total_templates_agg) %>%
    distinct()


  NT_by_patient <- tcr_raw %>%
    select(indv_label, rearrangement, productive_frequency) %>%
    pivot_wider(names_from = indv_label, values_from = productive_frequency, values_fill = 0, values_fn = sum)

  AA_by_patient <- tcr_raw %>%
    select(indv_label, amino_acid, productive_frequency) %>%
    pivot_wider(names_from = indv_label, values_from = productive_frequency, values_fill = 0, values_fn = sum)
  bioid_by_patient <- tcr_raw %>%
    select(indv_label, bio_identity, productive_frequency) %>%
    pivot_wider(names_from = indv_label, values_from = productive_frequency, values_fill = 0, values_fn = sum)

  paired_patients <- expand.grid(pt_x = unique(tcr_raw$indv_label), pt_y = unique(tcr_raw$indv_label)) %>%
    filter(pt_x != pt_y)

  list_of_paired_patient_nt <- apply(paired_patients, 1, function(row) {
    NT_by_patient %>%
      select(all_of(c(row[["pt_x"]], row[["pt_y"]])))
  })
  list_of_paired_patient_aa <- apply(paired_patients, 1, function(row) {
    AA_by_patient %>%
      select(all_of(c(row[["pt_x"]], row[["pt_y"]])))
  })
  list_of_paired_patient_bioid <- apply(paired_patients, 1, function(row) {
    bioid_by_patient %>%
      select(all_of(c(row[["pt_x"]], row[["pt_y"]])))
  })



  num_overlap_clones2 <- function(df) {
    # count rows where both columns are greater than zero
    nrow(df[df[, 1] > 0 & df[, 2] > 0, ])
  }

  num_shared_clone <- list_of_paired_patient_nt %>%
    map(~ num_overlap_clones2(.)) %>%
    unlist(.)

  num_shared_aa <- list_of_paired_patient_aa %>%
    map(~ num_overlap_clones2(.)) %>%
    unlist(.)

  num_shared_bioid <- list_of_paired_patient_bioid %>%
    map(~ num_overlap_clones2(.)) %>%
    unlist(.)

  paired_patients$num_shared_clones <- num_shared_clone
  paired_patients$num_shared_aa <- num_shared_aa
  paired_patients$num_shared_bioid <- num_shared_bioid

  shared_clones <- paired_patients

  ## to draw the venn diagram, we need to create a different format
  ## the format is a list of vector, each vector containing the clone names present in each patient

  # We removed the graft blood
  indvs <- unique(tcr_raw$indv_label)
  clones_per_pt <- lapply(setNames(indvs, indvs), function(x) {
    tcr_raw %>%
      filter(indv_label == x) %>%
      filter(tissue != "graft_blood") %>%
      filter(productive_frequency > 0) %>%
      pull(rearrangement) %>%
      unique()
  })
  aa_per_pt <- lapply(setNames(indvs, indvs), function(x) {
    tcr_raw %>%
      filter(indv_label == x) %>%
      filter(tissue != "graft_blood") %>%
      filter(productive_frequency > 0) %>%
      pull(amino_acid) %>%
      unique()
  })
  bioid_per_pt <- lapply(setNames(indvs, indvs), function(x) {
    tcr_raw %>%
      filter(indv_label == x) %>%
      filter(tissue != "graft_blood") %>%
      filter(productive_frequency > 0) %>%
      pull(bio_identity) %>%
      unique()
  })
  # this is the result from the AB-correlations.py script described above
  new_tcr_corrs <- read.csv(sep = "\t", "data/2022-08-29-tcr-correlations.tsv", header = FALSE, col.names = strsplit("fileA\tfileB\tsizeA\tsizeB\tnorm_sizeA\tnorm_sizeB\tJSD\tjsd_nt_norm\tjsd_aa_raw\tjsd_aa_norm\tmorisita_nt_raw\toverlapping_clones\toverlapping_aa", split = "\t")[[1]], na.strings = c("nan"))
  # do a double join to get all (non-unique combinations)
  tcr_corrs <- rbind(
    new_tcr_corrs %>% mutate(pair_is_unique = TRUE),
    new_tcr_corrs %>%
      filter(fileA != fileB) %>%
      mutate(
        fileX = fileA, fileA = fileB, fileB = fileX,
        sizeX = sizeA, sizeA = sizeB, sizeB = sizeX,
        norm_sizeX = norm_sizeA, norm_sizeA = norm_sizeB, norm_sizeB = norm_sizeX,
        pair_is_unique = FALSE
      ) %>%
      select(-fileX, -sizeX, -norm_sizeX)
  ) %>%
    rename(sample1 = fileA, sample2 = fileB) %>%
    distinct()
  tissue_annotations <- read.csv("data/updated_colors_dendrograms.csv")

  save(
    list = c(
      "tcr_sample_metadata",
      "tcr_corrs",
      "tissue_annotations",
      "clones_per_pt",
      "aa_per_pt",
      "bioid_per_pt",
      "tcr_raw"
    ),
    file = tcr_objects_path
  )
} else {
  load(tcr_objects_path)
}
```
