==> 20-exploratory-trees.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
```{r 20-trees-1, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, fig.width = 10, fig.height = 7)
library(tidyverse)
library(patchwork) # for combining figures
library(gplots) # for heatmap.2
library(ggtree) # for trees
library(ape) # for as.phylo
if (interactive()) load("tcr_objects.Rdata")
load("GTEx/gtex_objects.Rdata")
figs_dir <- "figures/exploratory_trees/"
dir.create(figs_dir, showWarnings = FALSE, recursive = TRUE)
```




## Plot Trees of all Human transplant patients
The following code creates plots showing the high-level relationships between the profiles of all the samples.  We show 4 sets of data:

- `just_patients`:  These are the 7 patients A-G described in the paper, without the 3 comparator patients H-J.
- `all`: All human and mice samples
- `all_no_donor_mice`: Same as `all`, but without the mouse donor Tcells.
- `all_no_donor_no_healthy_control_mice`: Same as `all`, but without the mouse donor Tcells or the Balb-c healthy comparator mice.


```{r 20-trees-2 }
for (batch in c("just_patients", "just_humans", "all", "all_no_donor_mice", "all_no_donor_no_healthy_control_mice")) {
  # batch="just_patients"
  if (batch == "just_patients") {
    these_samples <- tcr_sample_metadata %>% filter(org == "Human" & (gvhd == "GvHD" | indv_type == "Donor"))
  } else if (batch == "just_humans") {
    these_samples <- tcr_sample_metadata %>% filter(org == "Human")
  } else if (batch == "all") {
    these_samples <- tcr_sample_metadata
  } else if (batch == "all_no_donor_mice") {
    these_samples <- tcr_sample_metadata %>% filter(samp_type != "Donor")
  } else if (batch == "all_no_donor_no_healthy_control_mice") {
    these_samples <- tcr_sample_metadata %>% filter(samp_type != "Donor" & samp_type != "Healthy")
  }
  n_indv <- n_distinct(these_samples$indv_label)
  print(str_glue("making '{batch}' plots: {n_distinct(these_samples$sample_name)} samples from {n_indv} individuals"))
  for (metri in c("JSD")) {
    this_corr_data <- tcr_corrs %>%
      filter(sample1 %in% these_samples$sample_name) %>%
      filter(sample2 %in% these_samples$sample_name) %>%
      select(sample1, sample2, all_of(metri))
    this_corr_data$thismetric <- this_corr_data[, metri]
    this_corr_data[, metri] <- NULL
    tcr_agg_individuals <- this_corr_data %>%
      arrange(sample1, sample2) %>%
      pivot_wider(names_from = "sample2", values_from = thismetric) %>%
      column_to_rownames("sample1") %>%
      as.matrix()


    dend_ <- hclust(as.dist(tcr_agg_individuals), method = "single") %>% as.dendrogram()
    samples_ordered <- hclust(as.dist(tcr_agg_individuals), method = "single")$labels[c(hclust(as.dist(tcr_agg_individuals), method = "single")$order)]

    # get the clean, ordered sampleids
    samples_ordered_df <- left_join(data.frame(sample_name = samples_ordered), tcr_sample_metadata, by = "sample_name")
    if (!all(samples_ordered == samples_ordered_df$sample_name)) {
      stop("Error assigning clean name to ordered tissues")
    }
    samples_ordered_clean <- samples_ordered_df %>% pull(indv_tissue_label)

    dendextend::labels_colors(dend_) <- data.frame(Tissue = labels(dend_)) %>%
      left_join(., tissue_key) %>%
      pull(col)



    tmp <- ape::as.phylo(dend_)
    tmp$edge.length <- tmp$edge.length * 2
    tmpheight <- sort(stats::cophenetic(dend_), decreasing = TRUE)[1]

    ugly_specific_tissue_key <- this_corr_data %>%
      left_join(tcr_sample_metadata, by = c("sample1" = "sample_name")) %>%
      select(sample1, tissue, indv_label) %>%
      distinct()

    tick_pos <- seq(0, 1, .1)
    ticks <- data.frame(
      xorig = tick_pos, x = tick_pos - (1 - tmpheight),
      lab = rev(tick_pos), yend = rep(.3, length(tick_pos))
    )

    thesecolors <- rep(RColorBrewer::brewer.pal(n = 12, name = "Paired"), 4)[1:n_indv]
    names(thesecolors) <- unique(ugly_specific_tissue_key$indv_label)

    p_tcr_all <- ggtree(tmp) %<+% ugly_specific_tissue_key + xlim(-.05, 1.25) +
      geom_tiplab(aes(color = indv_label, label = gsub(" TCRB", "", gsub("_", " ", label))),
        offset = 0.01, # (1-max(tmp$edge.length)) + .01,
        align = FALSE, linetype = NA
      ) +
      geom_tippoint() +
      scale_linetype(guide = "none") +
      labs(color = "Patient") +
      scale_color_manual(values = thesecolors, guide = "none") +
      geom_segment(y = 0, yend = 0, x = min(ticks$x), xend = max(ticks$x)) +
      annotate(geom = "text", x = ticks$x, y = -1, label = ticks$lab) +
      geom_segment(data = ticks, aes(y = -yend, yend = 0, x = x, xend = x)) +
      theme(legend.position = c(.2, .95))

    # This nonsense adds the grouping label bars for the just_patients set.  Probably easier to just do in illustrator
    # if (metri=="JSD" & batch == "just_patients"){
    #   p_tcr_all <-  p_tcr_all +
    #     geom_cladelab(node=94, label=names(thesecolors)[1], angle=270, align=F, geom='text', textcolor=thesecolors[1], offset=.2, vjust=-.1, barcolor = thesecolors[1]) +
    #     geom_cladelab(node=98, label=names(thesecolors)[2], angle=270, align=F, geom='text', textcolor=thesecolors[2], offset=.2, vjust=-.1, barcolor = thesecolors[2]) +
    #     geom_cladelab(node=91, label=names(thesecolors)[3], angle=270, align=F, geom='text', textcolor=thesecolors[3], offset=.2, vjust=-.1, barcolor = thesecolors[3]) +
    #     geom_cladelab(node=96, label=names(thesecolors)[4], angle=270, align=F, geom='text', textcolor=thesecolors[4], offset=.2, vjust=-.1, barcolor = thesecolors[4]) +
    #     geom_cladelab(node=97, label=names(thesecolors)[5], angle=270, align=F, geom='text', textcolor=thesecolors[5], offset=.2, vjust=-.1, barcolor = thesecolors[5]) +
    #     geom_cladelab(node=86, label=names(thesecolors)[6], angle=270, align=F, geom='text', textcolor=thesecolors[6], offset=.2, vjust=-.1, barcolor = thesecolors[6]) +
    #     geom_cladelab(node=89, label=names(thesecolors)[7], angle=270, align=F, geom='text', textcolor=thesecolors[7], offset=.2, vjust=-.1, barcolor = thesecolors[7])
    # }
    ggsave(p_tcr_all, height = 3 + n_indv, width = 9, filename = file.path(figs_dir, str_glue("{metri}_TCR_{batch}_tree.pdf")))


    pmat_tcr_all <- ggplot(
      this_corr_data %>%
        mutate(
          sample1 = factor(sample1, levels = samples_ordered, ordered = TRUE),
          sample2 = factor(sample2, levels = samples_ordered, ordered = TRUE),
          lab = round(thismetric, 2),
        ), aes(x = sample1, y = sample2, fill = thismetric)
    ) +
      geom_tile() +
      coord_fixed(ratio = 1) +
      geom_text(aes(label = lab), color = "grey50", size = 1) +
      scale_fill_viridis_c(option = "A") +
      guides(fill = guide_colourbar(direction = "horizontal")) +
      scale_y_discrete(position = "right", expand = c(0, 0), labels = samples_ordered_clean) +
      scale_x_discrete(expand = c(0, 0), labels = samples_ordered_clean) +
      labs(subtitle = metri) +
      theme(axis.text.x = element_text(angle = 315, hjust = 0, size = 9), legend.position = "bottom")

    ggsave(pmat_tcr_all, width = 5 + n_indv, height = 5 + n_indv, filename = file.path(figs_dir, str_glue("{metri}_TCR_{batch}_heatmap.pdf")))
  }
}
```



```{r 20-trees-3 }
objects_for_saving <- c(
  "jsd_tcr_agg_individuals",
  "p_tcr_all",
  "pmat_tcr_all",
  "thesecolors"
)
for (ob in objects_for_saving) {
  if (!ob %in% ls()) {
    print(paste0("Warning: ", ob, " not in environment"))
    objects_for_saving <- objects_for_saving[objects_for_saving != ob]
  }
}
save(list = objects_for_saving, file = "tcr_graphic_objects.RData")
```
 
 
 ## Generate cross-patient tissue similarity plots for the GTEx transcription data

```{r 20-trees-4 }

for (metri in c("jsd")) {
  tick_pos <- seq(0, 1, .1)
  if (metri == "jsd") {
    this_ob <- jsd_agg_dist_corr
    tmpheight <- sort(stats::cophenetic(this_ob[[1]]), decreasing = TRUE)[1]
    ticks <- data.frame(
      xorig = tick_pos, x = tick_pos - (1 - tmpheight),
      lab = rev(tick_pos), yend = rep(.2, length(tick_pos))
    )
    ticks <- ticks %>% filter(xorig >= min(1 - this_ob[[2]]))
  } else {
    this_ob <- spear_agg_dist_corr
    tmpheight <- sort(stats::cophenetic(this_ob[[1]]), decreasing = TRUE)[1]
    ticks <- data.frame(
      xorig = tick_pos, x = tick_pos - (1 - tmpheight),
      lab = rev(tick_pos), yend = rep(.2, length(tick_pos))
    )
    ticks <- ticks %>% filter(xorig >= min(this_ob[[2]]))
  }
  tmp <- ape::as.phylo(this_ob[[1]])
  tmp$edge.length <- tmp$edge.length * 2
  # need to do this in cases lacking single root


  tmp_anno <- data.frame(sample_name = rownames(this_ob[[2]])) %>%
    mutate(
      patient = gsub("(.*?) - .*", "\\1", sample_name),
      tiss = gsub("(.*?) - (.*)", "\\2", sample_name),
    )
  (p <- ggtree(tmp) %<+% tmp_anno + xlim(min(ticks$x), max(ticks$x) * 3) +
    geom_tiplab(
      offset = 0.05, # (1-max(tmp$edge.length)) + .01,
      align = T, linetype = NA, size = 3,
    ) +
    geom_tippoint(position = position_nudge(x = 0.025), aes(color = patient, shape = tiss), size = 3) +
    labs(
      title = "GTEx",
      subtitle = str_glue("7 patients, distances based on {metri}")
    ) +
    scale_linetype(guide = "none") +
    scale_shape_manual(values = 1:13) +
    scale_color_brewer(palette = "Paired") +
    geom_segment(y = 0, yend = 0, x = min(ticks$x), xend = max(ticks$x)) +
    theme(legend.position = c(.2, .6)) +
    annotate(geom = "text", x = ticks$x, y = -.5, label = ticks$lab) +
    geom_segment(data = ticks, aes(y = -yend, yend = 0, x = x, xend = x))
  )


  ggsave(
    p,
    width = 8, height = 18, file = file.path(figs_dir, str_glue("GTEx_samples_alltogether_{metri}.pdf"))
  )
}
```
