==> 24-figure4-figure6-S19.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

## Venn 

```{R 24-figure4-1 }
if (interactive()) load("tcr_objects.Rdata")
fig4_figs_path <- file.path("figures", "fig4")
dir.create(fig4_figs_path, showWarnings = FALSE)


(V <- venn::venn(clones_per_pt[paste("Patient", LETTERS[1:7])], ilab = TRUE, zcolor = "style", opacity = .2, ggplot = TRUE))



ggsave(
  plot = V, height = 6, width = 6, dpi = 200,
  filename = file.path(fig4_figs_path, "fig4_venn_plot_7.pdf"),
  useDingbats = FALSE
)


(Vaa <- venn::venn(aa_per_pt[paste("Patient", LETTERS[1:7])], ilab = TRUE, zcolor = "style", opacity = .2, ggplot = TRUE))
ggsave(
  plot = Vaa, height = 6, width = 6,
  filename = file.path(fig4_figs_path, "fig4_venn_plot_7_aa.pdf"),
  useDingbats = FALSE
)

(Vbioid <- venn::venn(bioid_per_pt[paste("Patient", LETTERS[1:7])], ilab = TRUE, zcolor = "style", opacity = .2, ggplot = TRUE))
ggsave(
  plot = Vbioid, height = 6, width = 6,
  filename = file.path(fig4_figs_path, "fig4_venn_plot_7_bioid.pdf"),
  useDingbats = FALSE
)



# write.csv(paired_patients, "overlapping_clones_com.csv", row.names = F)

Vaa + Vbioid
```



```{R 24-figure4-2 }
# make one for each set because venns can only handle so many
for (thisrep in c("rep1", "rep2", "rep3")) {
  Vmouse <- venn::venn(clones_per_pt[grep(thisrep, names(clones_per_pt))], ilab = TRUE, zcolor = "style", opacity = .2, ggplot = TRUE)

  ggsave(
    plot = Vmouse, height = 6, width = 6,
    filename = file.path(fig4_figs_path, paste("fig4_venn_plot_7_mouse", thisrep, ".pdf")),
    useDingbats = FALSE
  )
}

Vmouse_donor <- venn::venn(aa_per_pt[grep("Donor", names(aa_per_pt))], ilab = TRUE, zcolor = "style", opacity = .2, ggplot = TRUE)

ggsave(
  plot = Vmouse_donor, height = 6, width = 6,
  filename = file.path("figures", "fig6", "6D-mouse_donor_venn_plot_AA.pdf"),
  useDingbats = FALSE
)
```


## GLIPH2 Analysis

Data files for GLIPH2 were exported by subsetting the mouse and human samples from the in-frame tcr data.


GLIPH2 was run with the parameters specified in `GLIPH2/test_parameter.txt` for human samples, and `GLIPH2/test_parameter_mouse.txt` for the mouse samples. The results were saved to `gliph`.  If you would like access to the raw GLIPH2 output prior to publication, please send an email to <dewolfs@mskcc.org>.

The human GLIPH2 data were processed with `GLIPH2/overlap_score_calculations.py`. Mouse GLIPH2 data were processed similarly with `GLIPH2/overlap_score_calculations_mouse.py`, which includes code for figure S10.
