==> 30-figureS2.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# Supplementary figures
## S2

Analysis of diversity metrics across clinical variables: these are the plots for supplemental figure 2.
```{R 30-figureS2-1 }
if (interactive()) load("./tcr_objects.Rdata")
S2_figs_dir <- file.path("figures", "S2")
dir.create(S2_figs_dir, showWarnings = FALSE)
figures_path <- file.path("figures", "S2")

s2_theme <- theme_classic() + theme(
  axis.text = element_text(size = 15, color = "black"),
  axis.text.y = element_text(size = 12, color = "black"),
  legend.text = element_text(size = 12, color = "black"), legend.title = element_text(size = 15, color = "black"), axis.title = element_text(size = 15, color = "black", face = "bold"),
  axis.text.x = element_text(size = 12, hjust = 0, color = "black", angle = -30)
)




s2diversity <- tcr_sample_metadata %>%
  select(sample_name:indv_tissue_label, -sample_amount_ng) %>%
  left_join(read.csv("tables/st4.csv") %>% rename(
    sample_name = Sample.name,
    productive_templates = productive.templates,
    max_productive_frequency = Max.frequency,
    len = unique.productive.clones
  )) %>%
  left_join(read.csv("data/patient_attributes.csv", na.strings = c("n/a")) %>%
    mutate(subject = paste("Patient", subject)), by = c("indv_label" = "subject")) %>%
  left_join(tissue_annotations, by = "tissue")

jitter2 <- position_jitter(width = 0.15, height = 0.0, seed = 12345)

GI <- s2diversity %>%
  filter(!tissue_label %in% c("PBSCs", "Bone Marrow", "pre-tx PBMCs", "Heart", "Kidney", "mLN")) %>%
  filter(tissue != "Graft/Blood") %>%
  filter(org == "Human") %>%
  filter("" != S2_groups)
GI_GVHD <- GI %>%
  filter(CMV.reactivation != "n/a")


# these are the plots for clonality
basep <- ggplot(GI, aes(x = reorder(S2_groups, anatomic_order)))

# this plot colors by allograft source
clonality_graft <- basep +
  geom_point(aes(y = clonality, fill = allograft), shape = 21, size = 5, stroke = 1, alpha = .8, position = jitter2) +
  s2_theme +
  xlab("") + scale_fill_brewer(palette = "Set1")

# this plot colors by allograft source
slope_graft <- basep +
  geom_point(aes(y = slope, fill = allograft), size = 5, stroke = 1, alpha = .8, position = jitter2, shape = 21) +
  s2_theme +
  xlab("") + scale_fill_brewer(palette = "Set1")

S2_A <- clonality_graft + slope_graft + plot_layout(guides = "collect")

S2_A

check_group_stats <- function(df, xcol, ycol) {
  thisdf <- df %>%
    rename(
      "x" = xcol,
      "y" = ycol
    ) %>%
    as.data.frame() %>%
    filter(!is.na(x))

  test_output <- lmerTest::lmer(y ~ x + tissue + (1 | indv_label), data = thisdf)
  test_output_summary <- summary(test_output)
  # slice gets rid of intercept
  sig_test <- test_output_summary$coefficients %>%
    as.data.frame() %>%
    slice(-1) %>%
    mutate(p.adj = p.adjust(`Pr(>|t|)`)) %>%
    filter(p.adj <= .05)

  if (nrow(sig_test) > 0) {
    stop(paste("significant association between", xcol, "and", ycol, "found!"))
  }
}
check_group_stats(df = GI, xcol = "allograft", ycol = "clonality")
check_group_stats(df = GI, xcol = "allograft", ycol = "slope")



# this plot colors by CMV
cc <- ggplot(GI_GVHD, aes(x = reorder(S2_groups, anatomic_order)))

clonality_CMV <- cc +
  geom_point(aes(y = clonality, fill = CMV.reactivation), size = 5, stroke = 1, alpha = .8, position = jitter2, shape = 21) +
  s2_theme +
  xlab("") + scale_fill_brewer(palette = "Set1")

slope_CMV <- cc +
  geom_point(aes(y = slope, fill = CMV.reactivation), size = 5, stroke = 1, alpha = .8, position = jitter2, shape = 21) +
  s2_theme +
  xlab("") + scale_fill_brewer(palette = "Set1")

S2_B <- clonality_CMV + slope_CMV + plot_layout(guides = "collect")

check_group_stats(df = GI_GVHD, xcol = "CMV.reactivation", ycol = "clonality")
check_group_stats(df = GI_GVHD, xcol = "CMV.reactivation", ycol = "slope")



# this plot colors by grade of GVHD peri-autopsy (acute GVHD)
clonality_GVHD <- cc +
  geom_point(aes(y = clonality, fill = as.factor(GVHD_severity)), size = 5, stroke = 1, alpha = .8, position = jitter2, shape = 21) + s2_theme + labs(fill = "GVHD \nseverity") + xlab("") + scale_fill_brewer(palette = "Set1")

# this plot colors by grade of GVHD peri-autopsy (acute GVHD)
slope_GVHD <- cc +
  geom_point(aes(y = slope, fill = as.factor(GVHD_severity)), size = 5, stroke = 1, alpha = .8, position = jitter2, shape = 21) +
  s2_theme + scale_fill_brewer(palette = "Set1") + xlab("") + labs(fill = "GVHD \nseverity")

S2_C <- clonality_GVHD + slope_GVHD + plot_layout(guides = "collect")

S2_C


check_group_stats(df = GI_GVHD, xcol = "GVHD_severity", ycol = "clonality")
check_group_stats(df = GI_GVHD, xcol = "GVHD_severity", ycol = "slope")


check_group_stats(df = GI, xcol = "gvhd", ycol = "clonality")
check_group_stats(df = GI, xcol = "gvhd", ycol = "slope")



# this plot colors by grade of age of patient
clonality_age <- cc +
  geom_point(aes(y = clonality, fill = as.factor(Pt_age_cat)), size = 5, stroke = 1, alpha = .8, position = jitter2, shape = 21) + s2_theme + labs(fill = "Pt_age_cat") + xlab("") + scale_fill_brewer(palette = "Set2")


# this plot colors by age of patient
slope_age <- cc +
  geom_point(aes(y = slope, fill = as.factor(Pt_age_cat)), size = 5, stroke = 1, alpha = .8, position = jitter2, shape = 21) +
  s2_theme + scale_fill_brewer(palette = "Set2") + xlab("") + labs(fill = "Pt_age_cat")

S2_age <- clonality_age + slope_age + plot_layout(guides = "collect")

check_group_stats(df = GI_GVHD, xcol = "Pt_age_cat", ycol = "clonality")
check_group_stats(df = GI_GVHD, xcol = "Pt_age_cat", ycol = "slope")





# this plot colors by grade of age of donor
clonality_donor <- cc +
  geom_point(aes(y = clonality, fill = as.factor(Donor_age_cat)), size = 5, stroke = 1, alpha = .8, position = jitter2, shape = 21) + s2_theme + labs(fill = "Donor_age_cat") + xlab("") + scale_fill_brewer(palette = "Set2")


# this plot colors by age of donor
slope_donor <- cc +
  geom_point(aes(y = slope, fill = as.factor(Donor_age_cat)), size = 5, stroke = 1, alpha = .8, position = jitter2, shape = 21) +
  s2_theme + scale_fill_brewer(palette = "Set2") + xlab("") + labs(fill = "Donor_age_cat")

S2_donor <- clonality_donor + slope_donor + plot_layout(guides = "collect")

check_group_stats(df = GI_GVHD, xcol = "Donor_age_cat", ycol = "clonality")
check_group_stats(df = GI_GVHD, xcol = "Donor_age_cat", ycol = "slope")



ggsave(plot = S2_A, height = 5, width = 14, filename = file.path(figures_path, "S2A_by_graft.pdf"))
ggsave(plot = S2_B, height = 5, width = 14, filename = file.path(figures_path, "S2B_by_CMV.pdf"))
ggsave(plot = S2_C, height = 5, width = 14, filename = file.path(figures_path, "S2C_by_GVHD.pdf"))
ggsave(plot = S2_age, height = 5, width = 14, filename = file.path(figures_path, "S2D_by_pt_age.pdf"))
ggsave(plot = S2_donor, height = 5, width = 14, filename = file.path(figures_path, "S2E_by_donor_age.pdf"))

```
