==> 39-figure-S3-S8-S11.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
```{r setup-39, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, fig.width = 10, fig.height = 7)
library(tidyverse)
library(patchwork) # for combining figures
library(RColorBrewer) # color palates
library(ggtree) # for trees
library(phyloseq)
library(dendextend) # for tanglegrams
library(ggrepel)
library(ggpubr)
library(rstatix)


theme_set(theme_classic())
if (interactive()) load("tcr_objects.Rdata")
if (interactive()) load("GTEx/gtex_objects.RData")
boxdir <- file.path("figures", "boxplot_comparisons")
dir.create(boxdir, showWarnings = FALSE)
figures_path <- file.path("figures", "boxplot_comparisons")
tcr_corrs_anno <- tcr_corrs %>%
  left_join(tcr_sample_metadata %>% select(sample_name:species), by = c("sample1" = "sample_name")) %>%
  left_join(tcr_sample_metadata %>% select(sample_name:species), by = c("sample2" = "sample_name"))
```

# Validating JSD trees

## Quantifying within vs between system distances

These figures were requested by a helpful reviewer who wished to see the between -vs within system distances quantified. For reference, the following figure shows the distances between tissues by the median GTEx summarized data:

```{r 39-figureS11-1 }
comp_df <- jsd_individuals_dist_list[["Tissues_in_TCR"]][[2]] %>%
  as_tibble(rownames = "rawA") %>%
  pivot_longer(cols = -rawA, names_to = "rawB") %>%
  mutate(A = pmin(rawA, rawB), B = pmax(rawA, rawB)) %>%
  select(A, B, value) %>%
  distinct() %>%
  filter(A != B) %>%
  left_join(specific_tissue_key, by = c("A" = "Tissue")) %>%
  left_join(specific_tissue_key, by = c("B" = "Tissue")) %>%
  mutate(grp = ifelse(TissueSystem.x == TissueSystem.y, "Within", "Between"))

ggplot(comp_df, aes(x = TissueSystem.y, color = grp, group = interaction(TissueSystem.y, grp), y = value)) +
  geom_boxplot(outlier.colour = NA) +
  geom_point(position = position_jitterdodge(jitter.height = 0, jitter.width = .1)) +
  labs(x = "Tissue System", y = "JSD", title = "GTEx median gene expression JSD between vs within Anatomic Systems")
```

Now, for the TCR data.  First, the per-patient boxplots showing the JSDs by tissue system, with stats for cases where multiple tissues within a system are available.

```{r 39-figureS11-2 }

### Per patient
for (pat in unique(tcr_sample_metadata$indv_label)) {
  print(str_glue("Processing {pat}"))
  # pat="Patient A"
  if (startsWith(pat, "Donor")) next
  pt_comp_df <- tcr_corrs_anno %>%
    filter(indv_label.x == pat & indv_label.y == pat) %>%
    rename(A = tissue.x, B = tissue.y, value = JSD) %>%
    select(A, B, value) %>%
    distinct() %>%
    filter(A != B) %>%
    left_join(tissue_key %>% select(TissueTCR, TissueSystem), by = c("A" = "TissueTCR")) %>%
    left_join(tissue_key %>% select(TissueTCR, TissueSystem), by = c("B" = "TissueTCR")) %>%
    mutate(grp = ifelse(TissueSystem.x == TissueSystem.y, "Within", "Between"))

  p_ <- ggplot(pt_comp_df, aes(x = TissueSystem.y, color = grp, group = interaction(TissueSystem.y, grp), y = value)) +
    geom_boxplot(outlier.colour = NA) +
    geom_point(position = position_jitterdodge(jitter.height = 0, jitter.width = .1), alpha = .5) +
    ggplot(pt_comp_df %>% group_by(TissueSystem.y) %>% filter(n_distinct(grp) > 1), aes(x = TissueSystem.y, color = grp, group = interaction(TissueSystem.y, grp), y = value)) +
    geom_boxplot(outlier.colour = NA, alpha = .4) +
    geom_point(alpha = .5, position = position_jitterdodge(jitter.height = 0, jitter.width = .1)) +
    stat_compare_means() +
    plot_layout(nrow = 2)
  ggsave(p_, filename = file.path(boxdir, str_glue("{pat}_comparisons.png")), width = 9, height = 8, dpi = 200)
}
```

Then the same was calculated across all samples.  We calculated these separately for the mouse and the human samples.

```{r 39-figureS11-3 }


tcr_corrs_anno <- tcr_corrs %>%
  left_join(tcr_sample_metadata %>% select(sample_name:species), by = c("sample1" = "sample_name")) %>%
  left_join(tcr_sample_metadata %>% select(sample_name:species), by = c("sample2" = "sample_name"))

SI_tissues <- tissue_annotations[tissue_annotations$uc_group %in%  c("Small Intestine"), "tissue"]  
LI_tissues <- tissue_annotations[tissue_annotations$uc_group %in%  c("Large Intestine"), "tissue"]  
GI_tissues <- c(SI_tissues, LI_tissues)

pt_comp_df_all_redundant <- tcr_corrs_anno %>%
  filter(tissue.x != "graft_blood" & tissue.y != "graft_blood") %>%
  filter(indv_label.x == indv_label.y) %>%
  filter(tissue.x != tissue.y) %>%
  mutate(pt=indv_label.x, org=org.x, gvhd=gvhd.x, value=JSD) %>% 
  left_join(tissue_annotations %>% select(tissue, between_within_groups), by = c("tissue.x" = "tissue")) %>% 
  left_join(tissue_annotations %>% select(tissue, between_within_groups), by = c("tissue.y" = "tissue"))

#buckle up.  We are faced with the problem of how to plot grouped parwise comparisons, without duplicates, but retaining redundant comparisons across groups
# so, the first thing we need to do is double the data swapping the groupings to ensure that all comparisons are present in each group
# (see the rbind below)
# then, group by patient and by the group used for the x axis/facets ("between_within_groups.x"),
# then do our sorting pmin magic to reorder the A's and Bs 
# then distinct()-ify to remove the duplicated comparisons.
# If we didn't do the grouping by pt AND by the grouping variable, we end up with
# fewer pairs due to the alphabetic ordering of the sorted pairs.
pt_comp_df_all <- rbind(pt_comp_df_all_redundant, 
                                  pt_comp_df_all_redundant %>% 
                                    mutate(tmptmp = between_within_groups.x,
                                           between_within_groups.x = between_within_groups.y,
                                           between_within_groups.y = tmptmp) %>% 
                                    select(-tmptmp)) %>% 
  group_by(pt, between_within_groups.x) %>% 
  mutate(A = pmin(tissue.x, tissue.y), B = pmax(tissue.x, tissue.y)) %>%
  select(A, B, value, between_within_groups.x, between_within_groups.y, pt, org, gvhd)  %>% 
  distinct() %>% 
  ungroup() %>% 
  left_join(tissue_annotations %>% select(-between_within_groups), by = c("A" = "tissue")) %>%
  left_join(tissue_annotations %>% select(-between_within_groups), by = c("B" = "tissue")) %>%
  mutate(grp = ifelse(between_within_groups.x == between_within_groups.y, "Within", "Between")) 

# old way -- this retains only a single record of each tissue pairing
# this still gets used for the within/between simulation below
pt_comp_df_all_nonredundant <- tcr_corrs_anno %>%
  filter(indv_label.x == indv_label.y) %>%
  mutate(A = pmin(tissue.x, tissue.y), B = pmax(tissue.x, tissue.y), value = JSD, pt = indv_label.x, org = org.x, gvhd = gvhd.x) %>%
  select(A, B, value, pt, org, gvhd)  %>%
  filter(A != "graft_blood" & B != "graft_blood") %>%
  left_join(tissue_annotations, by = c("A" = "tissue")) %>%
  left_join(tissue_annotations, by = c("B" = "tissue")) %>%
  mutate(grp = ifelse(between_within_groups.x == between_within_groups.y, "Within", "Between")) %>%
  distinct() %>%
  filter(A != B)

ggplot(pt_comp_df_all %>% 
         filter(org== "Human") %>% 
         filter(gvhd== "GvHD") %>% 
         mutate(between_within_groups.x = ifelse(between_within_groups.x == "Spleen", "blood", between_within_groups.x)) %>% 
         filter(A %in% c("autopsy_blood",  "spleen", GI_tissues)) %>% 
         filter(B %in% c("autopsy_blood",  "spleen", GI_tissues)) %>% 
         filter(!(grepl("Intestine",  between_within_groups.x) & grp == "Between")), 
       aes(x = grp, color = grp, group = grp, y = value)) +
    geom_boxplot(outlier.colour = NA) +
    geom_point(position = position_jitterdodge(jitter.height = 0, jitter.width = .1)) +
    labs(x = "System", y = "JSD", color = "Comparisons") +
    facet_grid(gvhd ~ between_within_groups.x, scales = "free", space = "free_x") +
    scale_color_manual(values = c("#3949AB", "#64B5F6"), breaks = c("Between", "Within"), labels = c("Between systems", "Within systems")) +
    theme(legend.position = "bottom") +
    scale_x_discrete(position = "top") 
ggplot(pt_comp_df_all %>% filter(alt_grp != "NA") %>% filter(org == "Human")%>% filter(between_within_groups.x != between_within_groups.y), aes(x=alt_grp, y=value)) + geom_boxplot() + geom_point() +
  facet_wrap(~org) 

# pt_comp_df_all_alternative <- tcr_corrs_anno %>%
#   filter(indv_label.x == indv_label.y) %>%
#   filter(tissue.x != tissue.y) %>%
#   left_join(tissue_annotations, by = c("tissue.x" = "tissue")) %>%
#   left_join(tissue_annotations, by = c("tissue.y" = "tissue")) %>%
#   mutate(grp = ifelse(between_within_groups.x == between_within_groups.y, "Within", "Between"))
# 
# 
# ggplot(pt_comp_df_all_alternative %>% filter((org.x =="Mouse" & tissue.x == "spleen") | (org.x == "Human" & tissue.x=="autopsy_blood") ), aes(x=indv_label.x, y=JSD)) + geom_boxplot() + geom_jitter() + facet_wrap(~org.x, scales = "free") + theme(axis.text.x = element_text(angle=45, hjust=1))
# 
# ggplot(pt_comp_df_all_alternative %>% 
#          filter(tissue.x %in% c("blood", "autopsy_blood", "spleen")) %>%
#          mutate(grp = (org.x =="Mouse" & tissue.x == "spleen") | (org.x == "Human" & tissue.x=="autopsy_blood")  | (org.x =="Mouse" & tissue.y == "spleen") | (org.x == "Human" & tissue.y=="autopsy_blood") ), aes(x=indv_label.x, y=JSD, color=grp)) + geom_boxplot() + geom_jitter() + facet_wrap(~org.x, scales = "free") + theme(axis.text.x = element_text(angle=45, hjust=1))
# 
# invalid_tissues = c("graft_blood", "autopsy_marrow", "pretx_blood", )
# ggplot(pt_comp_df_all_alternative %>% 
#          filter(org.x == "Human") %>% 
#          filter(!tissue.x %in% invalid_tissues) %>% 
#          filter(!tissue.y %in% invalid_tissues) %>% 
#          mutate(grp =  ifelse((org.x == "Human" & tissue.x=="autopsy_blood")   | (org.x == "Human" & tissue.y=="autopsy_blood"), "Blood", "non-blood" )), aes(x=indv_label.x, y=JSD, color=grp)) + geom_boxplot() + geom_point(position=position_jitterdodge()) + facet_wrap(~org.x, scales = "free") + theme(axis.text.x = element_text(angle=45, hjust=1))

# tissue overlap
tmp1 <- pt_comp_df_all_alternative %>% 
  mutate(pct_overlap = overlapping_clones / sizeB) %>% 
  filter(org.x == "Human") %>% 
  filter(tissue.x == "autopsy_blood") %>% 
  filter(!tissue.y %in% invalid_tissues) %>% 
  ggplot(aes(x=indv_label.x, y=pct_overlap )) + 
   geom_boxplot() + geom_jitter() + scale_y_continuous(limits=c(0, 1)) + 
 theme(axis.text.x = element_text(angle=45, hjust=1)) +  labs(title="Human Percent of tissue overlap from blood")
# JSD
tmp2 <- pt_comp_df_all_alternative %>% 
  mutate(pct_overlap = overlapping_clones / sizeB) %>% 
  filter(org.x == "Human") %>% 
  filter(tissue.x == "autopsy_blood") %>% 
  filter(!tissue.y %in% invalid_tissues) %>% 
  ggplot(aes(x=indv_label.x, y=JSD )) + 
   geom_boxplot() + geom_jitter() + scale_y_continuous(limits=c(0, 1))  +
  labs(title="Human JSD from dlood")  + theme(axis.text.x = element_text(angle=45, hjust=1))



# same, but for mice
# tissue overlap
tmp3 <- pt_comp_df_all_alternative %>% 
  mutate(pct_overlap = overlapping_clones / sizeB) %>% 
  filter(org.x == "Mouse") %>% 
  filter(tissue.x == "spleen") %>% 
  filter(!tissue.y %in% invalid_tissues) %>% 
  ggplot(aes(x=indv_label.x, y=pct_overlap )) + 
   geom_boxplot() + geom_jitter() + scale_y_continuous(limits=c(0, 1)) + 
  labs(title="Mouse Percent of tissue overlap from blood")  + theme(axis.text.x = element_text(angle=45, hjust=1))
# JSD
tmp4 <- pt_comp_df_all_alternative %>% 
  mutate(pct_overlap = overlapping_clones / sizeB) %>% 
  filter(org.x == "Mouse") %>% 
  filter(tissue.x == "spleen") %>% 
  filter(!tissue.y %in% invalid_tissues) %>% 
  ggplot(aes(x=indv_label.x, y=JSD )) + 
   geom_boxplot() + geom_jitter() + scale_y_continuous(limits=c(0, 1)) + 
  labs(title="Mouse JSD from dlood") + theme(axis.text.x = element_text(angle=45, hjust=1))

ggsave(tmp1 +tmp2 + tmp3 + tmp4 + plot_layout(nrow=2), filename = "tmp_reviewer_comment_divergence_from_blood.pdf", width = 16, height = 16)




for (thisorg in c("Mouse", "Human")) {
  if (thisorg == "Mouse") {
    thisdata <- pt_comp_df_all %>%
      filter(between_within_groups.x != "") %>%
      filter(org == thisorg) %>%
      filter(gvhd == "GvHD")
  } else {
    thisdata <- pt_comp_df_all %>%
      filter(between_within_groups.x != "") %>%
      filter(org == thisorg) %>% 
      select(pt, A, B, grp, between_within_groups.x, between_within_groups.y, gvhd, value)
  }
  # find all the tissues with within-group comparisons
  stat.test <- thisdata %>%
    group_by(between_within_groups.x, gvhd) %>%
    filter(n_distinct(grp) > 1) %>%
    rstatix::kruskal_test(value ~ grp) %>%
    adjust_pvalue(method = "BH") %>%
    add_significance("p.adj",
      cutpoints = c(0, 1e-04, 0.001, 0.01, 0.05, 1),
      symbols = c("****", "***", "**", "*", "ns")
    ) %>%
    mutate(group1 = "Between", group2 = "Within") %>%
    add_xy_position(x = "grp", dodge = 0.8)


  p_between_within <- ggplot(thisdata, aes(x = grp, color = grp, group = grp, y = value)) +
    geom_boxplot(outlier.colour = NA) +
    geom_point(position = position_jitterdodge(jitter.height = 0, jitter.width = .1)) +
    labs(x = "System", y = "JSD", color = "Comparisons") +
    facet_grid(gvhd ~ between_within_groups.x, scales = "free", space = "free_x") +
    scale_color_manual(values = c("#3949AB", "#64B5F6"), breaks = c("Between", "Within"), labels = c("Between systems", "Within systems")) +
    theme(legend.position = "bottom") +
    scale_x_discrete(position = "top") +
    stat_pvalue_manual(
      stat.test,
      label = "{p.adj.signif}", tip.length = 0
    ) +
    theme(
      axis.ticks.x = element_blank(), axis.text.x = element_blank(),
    ) 

  ggsave(p_between_within, filename = file.path(boxdir, paste0("between_within_", thisorg, ".pdf")), width = 16, height = 8)
  ggsave(p_between_within +
    gridExtra::tableGrob(tissue_annotations %>% select(between_within_groups, tissue) %>% distinct() %>% arrange(desc(between_within_groups)),
      theme = gridExtra::ttheme_minimal(base_size = 7),
      rows = NULL
    ) +
    plot_layout(widths = c(8, 2)), filename = file.path(boxdir, paste0("between_within_", thisorg, "with_table.pdf")), width = 16, height = 8)
  
  # revision 2 figure
  # Note this doesn't have blood vs spleen 
  if (thisorg == "Human"){
    # ffor the human, limit analysis only to those patients with autopsy blood samples
    these_patients <- tcr_sample_metadata %>% filter(org=="Human" & gvhd=="GvHD" & grepl("autopsy_blood", sample_name)) %>% pull(indv_label)
    blood_comparisons <- list(
      c("Blood", "Large Intestine"),
      c("Blood", "Small Intestine"))
    spleen_comparisons <- list(
      c("Spleen", "Large Intestine"),
      c("Spleen", "Small Intestine"))
  } else{
    these_patients = unique(thisdata$pt)
     blood_comparisons <- list(
      c("Blood", "Small Intestine"))
    spleen_comparisons <- list(
      c("Spleen", "Small Intestine"))
 }
  thisdata_subset <- thisdata %>% 
    filter(pt %in% these_patients) %>% 
    mutate(newgroup = case_when(
      (A %in% c("autopsy_blood", "blood") & B %in% GI_tissues  & between_within_groups.x == "blood" ) ~ "Blood",
      (B %in% c("autopsy_blood", "blood") & A %in% GI_tissues  & between_within_groups.x == "blood" ) ~ "Blood",
      # (A == "spleen" & B %in% GI_tissues & between_within_groups.x == "Spleen" )  ~ "Spleen",
      # (B == "spleen" & A %in% GI_tissues & between_within_groups.x == "Spleen")  ~ "Spleen",
      (A %in%  LI_tissues & grp == "Within")  ~ "Large Intestine",
      (B %in%  LI_tissues & grp == "Within")  ~ "Large Intestine",
      (A %in%  SI_tissues & grp == "Within")  ~ "Small Intestine",
      (B %in%  SI_tissues & grp == "Within")  ~ "Small Intestine",
      TRUE ~ "Ignore"
    ) 
    ) %>% 
    filter(newgroup != "Ignore") %>% 
    filter(gvhd == "GvHD")
  


  p_between_within_subset <- ggplot(thisdata_subset, aes(x =newgroup,  y = value)) +
    geom_boxplot(outlier.colour = NA) +
    geom_jitter(height = 0, width = .1, color="black") +
    labs(x = "System", y = "JSD", color = "Comparisons") +
    theme(legend.position = "bottom") +
    stat_compare_means(comparisons = blood_comparisons , method = "wilcox.test")

  ggsave(p_between_within_subset, filename = file.path(boxdir, paste0("between_within_", thisorg, "_blood_GI_subset.pdf")), width = 16, height = 8)
  
  
  #######   Spleen
  thisdata_subset_spleen <- thisdata %>% 
    mutate(newgroup = case_when(
      (A == "spleen" & B %in% GI_tissues & between_within_groups.x == "Spleen" )  ~ "Spleen",
      (B == "spleen" & A %in% GI_tissues & between_within_groups.x == "Spleen")  ~ "Spleen",
      (A %in%  LI_tissues & grp == "Within")  ~ "Large Intestine",
      (B %in%  LI_tissues & grp == "Within")  ~ "Large Intestine",
      (A %in%  SI_tissues & grp == "Within")  ~ "Small Intestine",
      (B %in%  SI_tissues & grp == "Within")  ~ "Small Intestine",
      TRUE ~ "Ignore"
    ) 
    ) %>% 
    filter(newgroup != "Ignore") %>% 
    filter(gvhd == "GvHD")
  


  p_between_within_subset_spleen <- ggplot(thisdata_subset_spleen, aes(x =newgroup,  y = value)) +
    geom_boxplot(outlier.colour = NA) +
    geom_jitter(height = 0, width = .1, color="black") +
    labs(x = "System", y = "JSD", color = "Comparisons") +
    theme(legend.position = "bottom") +
    stat_compare_means(comparisons = spleen_comparisons )

  ggsave(p_between_within_subset_spleen, filename = file.path(boxdir, paste0("between_within_", thisorg, "_spleen_GI_subset.pdf")), width = 16, height = 8)

}
file.copy(
  file.path(boxdir, paste0("between_within_Human.pdf")),
  file.path(boxdir, paste0("S3E-between_within_Human.pdf"))
)
file.copy(
  file.path(boxdir, paste0("between_within_Mouse.pdf")),
  file.path(boxdir, paste0("S8C-between_within_Mouse.pdf"))
)
```

## Permuttaion tests
Next, we shuffled the classes for the systems with multiple tissues represented in the data (limiting to human samples for simplicity).  The probability weight of the sampling was set to mimic the comparisons at hand. For instance, we have 124 small-intestinal vs non-small-intestine pairs, and 19 within-small-intestine pairs, so a weight ratio or .867:.133.

```{r 39-figureS11-4 }
(weights <- pt_comp_df_all_nonredundant %>%
  filter(org == "Human") %>%
  group_by(plotgroup.x, org) %>%
  filter(n_distinct(grp) > 1) %>%
  mutate(plotgroup_n = n()) %>%
  group_by(plotgroup.x, org, grp) %>%
  dplyr::summarize(pn = plotgroup_n[1], n = n(), weight = n() / pn)
)
between_weight <- weights %>%
  filter(grp == "Between") %>%
  pull(weight) %>%
  mean()
```


The sampling was performed 1000 times, and plotted below.  



```{r 39-figureS11-5 }
tissues_for_comparisons <- pt_comp_df_all_nonredundant %>%
  filter(org == "Human") %>%
  filter(between_within_groups.x != "") %>%
  group_by(between_within_groups.x, org) %>%
  filter(n_distinct(grp) > 1) %>%
  pull(between_within_groups.x) %>%
  unique()

reps <- 1000

pt_comp_df_all_shuff <- map(1:reps, .f = function(x) {
  set.seed(x)
  pt_comp_df_all_nonredundant %>%
    filter(org == "Human") %>% # gvhd == "GvHD") %>%
    filter(between_within_groups.x != "") %>%
    filter(between_within_groups.x %in% tissues_for_comparisons) %>%
    group_by(between_within_groups.x, org) %>%
    mutate(
      shuffle = x,
      fauxgrp = sample(c("Within", "Between"), replace = TRUE, size = n(), prob = c(1 - between_weight, between_weight))
    )
}) %>% bind_rows()

# only plot the first 6 cause 100 is a bit excessive
p_inv_shuff <- ggplot(pt_comp_df_all_shuff %>% filter(shuffle < 7), aes(x = between_within_groups.x, color = fauxgrp, group = interaction(between_within_groups.x, fauxgrp), y = value)) +
  geom_boxplot(outlier.colour = NA) +
  geom_point(position = position_jitterdodge(jitter.height = 0, jitter.width = .1)) +
  stat_compare_means() +
  facet_wrap(~shuffle, ncol = 2) +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  labs(caption = "Each facet is a replicate of randomly assigning the 'Within'/'Between' grouping to comparisons relating to a anatomic system. \nFirst 6 replicates shown for clarity.")
ggsave(p_inv_shuff, filename = file.path(boxdir, "shuffles_between_within_first_six.pdf"), width = 16, height = 10)



shuff_summary <- pt_comp_df_all_shuff %>%
  group_by(between_within_groups.x, shuffle) %>%
  rstatix::kruskal_test(value ~ fauxgrp) %>%
  adjust_pvalue(method = "BH")

stat.test.allhuman <- pt_comp_df_all_nonredundant %>%
  filter(between_within_groups.x != "") %>%
  filter(org == thisorg) %>%
  group_by(between_within_groups.x) %>%
  filter(n_distinct(grp) > 1) %>%
  rstatix::kruskal_test(value ~ grp) %>%
  adjust_pvalue(method = "BH") %>%
  add_significance("p.adj",
    cutpoints = c(0, 1e-04, 0.001, 0.01, 0.05, 1),
    symbols = c("****", "***", "**", "*", "ns")
  ) %>%
  mutate(group1 = "Between", group2 = "Within") %>%
  add_xy_position(x = "grp", dodge = 0.8)


perm_test <- shuff_summary %>%
  left_join(stat.test.allhuman %>% select(between_within_groups.x, p.adj) %>% rename(true_p.adj = p.adj), by = "between_within_groups.x") %>%
  group_by(between_within_groups.x, true_p.adj) %>%
  summarize(n = n(), n_less_than_true = sum(p <= true_p.adj)) %>%
  mutate(permuted_p_text = str_glue("{n_less_than_true} / {n}"))

(p_shuff_summary <- ggplot(shuff_summary, aes(y = between_within_groups.x, group = between_within_groups.x, x = p)) +
  geom_violin(alpha = .2) +
  geom_jitter(width = 0, height = .1, alpha = .2, color = "black") +
  geom_point(data = stat.test.allhuman, aes(x = p.adj, y = between_within_groups.x, color = between_within_groups.x), shape = 8, size = 3, stroke = 2) +
  scale_color_brewer(palette = "Dark2", guide = "none") +
  geom_text(data = perm_test, aes(x = ifelse(true_p.adj > .1, true_p.adj, .1), y = between_within_groups.x, color = between_within_groups.x, label = permuted_p_text), nudge_y = -.25) +
  labs(x = "P-value", y = "System", title = "System Permutations", caption = str_wrap("Asterisks show adjusted p-value of true system labels; black points points and violin outline show non-adjusted p-values of permutated system labels. Colored text shows number of tests equal to or less than the adjusted p-value for 1000 permutations.", width = 70)) +
  scale_x_log10())


ggsave(p_shuff_summary, filename = file.path(boxdir, "S3F-shuffles_between_within_summary.pdf"), width = 5, height = 5)
```


Lastly, we compared the broad categories of intestinal vs non-intestinal, rather than by tissue system:

```{R 39-figureS11-6 }
sets_of_comparisons <- list(
  "intestinal" = c("Large Intestine", "Small Intestine"),
  "GI" = c("Large Intestine", "Small Intestine", "mLN", "Upper GI")
)


for (si in 1:length(sets_of_comparisons)) {

  # comaprison set
  s <- sets_of_comparisons[[si]]
  # name of comparison set
  sname <- names(sets_of_comparisons)[[si]]

  intestinal_comparison_data <- pt_comp_df_all_nonredundant %>%
    filter(gvhd == "GvHD") %>%
    filter(between_within_groups.x %in% s | between_within_groups.y %in% s) %>%
    mutate(intestinal = ifelse(between_within_groups.x %in% s & between_within_groups.y %in% s, "Within", "Between"))

  stat.test.intestinal <- intestinal_comparison_data %>%
    group_by(org, ) %>%
    rstatix::wilcox_test(value ~ intestinal) %>%
    adjust_pvalue(method = "BH") %>%
    add_xy_position(x = "intestinal", dodge = 0.8)

  p_between_within_intestine <- ggplot(intestinal_comparison_data, aes(x = intestinal, color = intestinal, group = intestinal, y = value)) +
    geom_boxplot(outlier.colour = NA) +
    geom_point(position = position_jitterdodge(jitter.height = 0, jitter.width = .1)) +
    labs(x = "System", y = "JSD", title = str_glue("JSD within {sname} tissues")) +
    facet_grid(org ~ ., scales = "free", space = "free_x") +
    scale_color_manual(values = c("#3949AB", "#64B5F6"), breaks = c("Between", "Within"), labels = c(str_glue("Between {sname} tissues"), str_glue("Within {sname} tissues")), guide = "none") +
    theme(legend.position = "bottom") +
    stat_pvalue_manual(stat.test.intestinal, label = "Wilcox adj. p {p}", tip.length = 0) +
    scale_y_continuous(limits = c(0, 1.1), expand = c(0, 0))

  #   stat_compare_means(comparisons = list(c( "non-intestinal", "intestinal")), label = "Wilcox adj. p {p.adj}", tip.length = 0, method = "wilcox.test")


  nonintestinal_weight <- intestinal_comparison_data %>%
    group_by(org) %>%
    mutate(plotgroup_n = n()) %>%
    group_by(org, intestinal) %>%
    dplyr::summarize(pn = plotgroup_n[1], n = n(), weight = n() / pn) %>%
    filter(intestinal == "Between") %>%
    pull(weight) %>%
    mean()

  pt_comp_df_all_shuff_intestinal <- map(1:1000, .f = function(x) {
    set.seed(x)
    intestinal_comparison_data %>%
      group_by(org) %>%
      mutate(shuffle = x, fauxgrp = sample(c("Between", "Within"), replace = TRUE, size = n(), prob = c(1 - nonintestinal_weight, nonintestinal_weight)))
  }) %>% bind_rows()

  pt_comp_df_all_shuff_intestinal_summary <- pt_comp_df_all_shuff_intestinal %>%
    group_by(org, shuffle) %>%
    rstatix::wilcox_test(value ~ fauxgrp) %>%
    adjust_pvalue(method = "BH")

  this_perm_test <- pt_comp_df_all_shuff_intestinal_summary %>%
    left_join(stat.test.intestinal %>% select(org, p.adj) %>% rename(true_p.adj = p.adj)) %>%
    group_by(org, true_p.adj) %>%
    summarize(n = n(), n_less_than_true = sum(p <= true_p.adj)) %>%
    mutate(permuted_p_text = str_glue("{n_less_than_true} / {n}"))

  p_shuff_intestinal_summary <- ggplot(pt_comp_df_all_shuff_intestinal_summary, aes(y = org, group = org, x = p)) +
    geom_boxplot(outlier.colour = NA, alpha = .2) +
    geom_jitter(width = 0, height = .1, alpha = .2, color = "black") +
    scale_y_discrete(limits = rev(c("Human", "Mouse"))) +
    geom_point(data = stat.test.intestinal, aes(x = p.adj, y = rev(org), color = org), shape = 8, size = 3, stroke = 2) +
    scale_color_brewer(palette = "Set2", guide = "none") +
    # geom_text(data=this_perm_test, aes(x=ifelse(true_p.adj > .1, true_p.adj, .1), y=org, color=org, label=permuted_p_text), nudge_y = -.25)  +
    geom_text(data = this_perm_test, aes(x = true_p.adj, y = rev(org), color = org, label = permuted_p_text), nudge_y = -.25, nudge_x = .001) +
    labs(x = "Adjusted P-value", y = "Organism", title = "Significance when shuffled", subtitle = paste0(sname, " vs non-", sname)) +
    scale_x_log10()



  ggsave(p_between_within_intestine + p_shuff_intestinal_summary + plot_layout(ncol = 1, heights = c(.8, .3)), filename = file.path(boxdir, str_glue("mouse_human_between_within_{sname}.pdf")), width = 4, height = 8)
}
file.copy(
  file.path(boxdir, str_glue("mouse_human_between_within_{sname}.pdf")),
  file.path(boxdir, str_glue("S8B-mouse_human_between_within_GI.pdf"))
)
```

## Raw data comparison plots


```{r 39-figureS11-7 }
# note the lack of the `distinct()` call here; we need all the (redundant) comparisons for this figure
set.seed(123)

thistcr_corrs_anno <- tcr_corrs_anno %>%
  filter(sample1 != "PtA_autopsy_marrow_TCRB") %>%
  filter(sample2 != "PtA_autopsy_marrow_TCRB")

for (i in unique(thistcr_corrs_anno$indv_label.x)) {
  tmp_p <- ggplot(thistcr_corrs_anno %>%
    filter(indv_label.x == indv_label.y) %>% filter(indv_label.x == i) %>%
    left_join(tissue_annotations, by = c("tissue.x" = "tissue")) %>%
    filter(tissue.x != tissue.y) %>%
    left_join(tissue_annotations, by = c("tissue.y" = "tissue")), aes(x = reorder(tissue.x, anatomic_order.x), y = JSD, shape = tissue.y, color = tissue.y)) +
    geom_jitter(width = .2, height = 0, size = 5) +
    scale_color_manual(values = tissue_annotations$new_col, breaks = tissue_annotations$tissue) +
    scale_shape_manual(values = tissue_annotations$shape, breaks = tissue_annotations$tissue) +
    labs(x = "Tissue", shape = "Tissue", subtitle = i, color = "Tissue") +
    theme_classic() +
    theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
    scale_y_continuous(limits = c(0, 1))

  if (startsWith(i, "Patient")) {
    ggsave(plot = tmp_p, filename = file.path(boxdir, str_glue("S3B-{i}-TCR_raw_comparisons.pdf")), width = 6, height = 6)
  } else {
    ggsave(plot = tmp_p, filename = file.path(boxdir, str_glue("TCR_raw_comparisons-{i}.pdf")), width = 6, height = 6)
  }
}

# formatting individual graph to focus on comparisons with blood (Figure R1) --> this is just for the letter
tmp_gvhd <- ggplot(tcr_corrs_anno %>%
  filter(indv_label.x == indv_label.y) %>%
  filter(indv_label.x %in% c("Patient A", "Patient D", "Patient F", "Patient G")) %>%
  left_join(tissue_annotations, by = c("tissue.x" = "tissue")) %>%
  filter(tissue.x != tissue.y) %>% filter(tissue.x == "autopsy_blood") %>%
  filter(tissue.y != "graft_blood") %>%
  filter(tissue.y != "autopsy_marrow") %>%
  filter(tissue.y != "spleen") %>%
  filter(tissue.y != "mLN") %>%
  filter(tissue.y != "heart") %>%
  filter(tissue.y != "kidney") %>%
  filter(tissue.y != "pretx_blood") %>%
  left_join(read.csv("data/patient_attributes.csv", na.strings = c("n/a")) %>%
    mutate(subject = paste("Patient", subject)), by = c("indv_label.x" = "subject")) %>%
  left_join(tissue_annotations, by = c("tissue.y" = "tissue")) %>%
  mutate(indv_label.x = reorder(indv_label.x, GVHD_severity)), aes(x = indv_label.x, GVHD_severity, y = JSD)) +
  geom_violin() +
  geom_jitter(aes(color = tissue_label_fig1.y), width = .2, height = 0, size = 3) +
  scale_color_manual(values = tissue_annotations$new_col, breaks = tissue_annotations$tissue_label_fig1) +
  scale_shape_manual(values = tissue_annotations$shape, breaks = tissue_annotations$tissue_label_fig1) +
  labs(x = "GVHD Patient", shape = "Tissue", color = "Tissue") +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_y_continuous(limits = c(0, 1))

tmp_gvhd

ggsave(plot = tmp_gvhd, height = 4, width = 5, filename = file.path(figures_path, "GVHD_seversity_JSD.pdf"))

# mean, median, standard deviation for blood/GVHD comparisons
data.frame(tmp_gvhd$data %>% group_by(sample1) %>%
  summarize(
    mean_JSD = mean(JSD),
    median_clones_per_patient = median(JSD),
    sd_JSD = sd(JSD),
  ))
```
## Heatmaps


## Summarizing Mouse TCR data

```{r 39-figureS11-8 }
tcr_corrs_anno_mouse <- tcr_corrs_anno %>%
  filter(pair_is_unique == TRUE) %>%
  filter(org.x == "Mouse") %>%
  filter(org.y == "Mouse") %>%
  filter(indv_label.x == indv_label.y) %>%
  select(JSD, tissue.x, tissue.y, gvhd.x, collection_day.x) %>%
  group_by(across(c(-JSD))) %>%
  mutate(
    n = n(),
    JSD_min = min(JSD),
    JSD_max = max(JSD),
    JSD_mean = mean(JSD),
    JSD_median = median(JSD),
    JSD_var = var(JSD),
  ) %>%
  select(tissue.x, tissue.y, everything()) %>%
  ungroup()

# To reorder the heatmap columns according to anatomic order, we want to add back
# in the redundant pairs we intentionally excluded above.  We do that by binding in
# the data with the reversed labels.
tcr_corrs_anno_mouse_summary <- bind_rows(
  tcr_corrs_anno_mouse %>% select(-JSD) %>% distinct(),
  tcr_corrs_anno_mouse %>% select(-JSD) %>% distinct() %>% rename(tmp = tissue.x) %>% rename("tissue.x" = "tissue.y", "tissue.y" = "tmp")
)

p_mouse_variance <- ggplot(tcr_corrs_anno_mouse %>%
  filter(gvhd.x == "GvHD") %>%
  filter(tissue.x != tissue.y) %>%
  filter(collection_day.x == 14), aes(x = reorder(interaction(tissue.x, tissue.y), desc(JSD_var)), y = JSD, color = as.character(n))) +
  scale_color_discrete() +
  labs(color = "N") +
  theme(
    legend.position = c(.31, .1),
    legend.direction = "horizontal",
    legend.background = element_rect(color = "gray80", fill = NA),
    # legend.background = element_blank() ,
    axis.text.x = element_blank(),
    axis.title.x = element_blank(),
  ) +
  geom_jitter(height = 0, alpha = .3, width = .2) +
  ggplot(tcr_corrs_anno_mouse_summary %>%
    filter(tissue.x != tissue.y) %>%
    filter(gvhd.x == "GvHD") %>% filter(collection_day.x == 14), aes(x = reorder(interaction(tissue.x, tissue.y), desc(JSD_var)), y = JSD_var)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  geom_point() +
  labs(y = "Variance in JSD", x = "") +
  plot_layout(nrow = 2) +
  plot_annotation(title = "JSD and JSD variance across mouse tissue pairs (day 14)")


ggsave(filename = file.path(boxdir, "mouse_JSD_variance.pdf"), plot = p_mouse_variance, width = 9, height = 6)


expl_figs_dir <- "figures/exploratory_trees/"
dir.create(expl_figs_dir, showWarnings = FALSE)


mouse_summary_plist <- list()
for (gr in c("GvHD", "noGvHD")) {
  for (cday in c(7, 14)) {
    this_corr_data <- tcr_corrs_anno_mouse_summary %>%
      filter(gvhd.x == gr) %>%
      filter(tissue.x != "Tcells", tissue.y != "Tcells")
    # No days for healthy mice, take all of them
    if (gr == "GvHD") {
      this_corr_data <- this_corr_data %>%
        filter(collection_day.x == cday)
    }
    tcr_agg_mice <- this_corr_data %>%
      rename(sample1 = tissue.x, sample2 = tissue.y) %>%
      select(sample1, sample2, JSD_mean) %>%
      distinct() %>%
      arrange(desc(sample1), desc(sample2)) %>%
      pivot_wider(names_from = "sample2", values_from = JSD_mean) %>%
      column_to_rownames("sample1") %>%
      as.matrix()
    # the arrange() makes it possible to convert to a proper lower right tri distance matrix
    dend_ <- hclust(as.dist(tcr_agg_mice), method = "single") %>% as.dendrogram()
    samples_ordered <- hclust(as.dist(tcr_agg_mice), method = "single")$labels[c(hclust(as.dist(tcr_agg_mice), method = "single")$order)]

    tcr_agg_mice_ordered <- tcr_agg_mice[samples_ordered, samples_ordered]
    # get the clean, ordered sampleids
    samples_ordered_df <- left_join(data.frame(tissue = samples_ordered), tissue_annotations, by = "tissue")
    if (!all(samples_ordered == samples_ordered_df$sample_name)) {
      stop("Error assigning clean name to ordered tissues")
    }
    samples_ordered_clean <- samples_ordered_df %>% pull(tissue)

    dendextend::labels_colors(dend_) <- data.frame(Tissue = labels(dend_)) %>%
      left_join(., tissue_key) %>%
      pull(col)



    tmp <- ape::as.phylo(dend_)
    tmp$edge.length <- tmp$edge.length * 2
    tmpheight <- sort(stats::cophenetic(dend_), decreasing = TRUE)[1]

    #
    tick_pos <- seq(0, 1, .1)
    ticks <- data.frame(
      xorig = tick_pos, x = tick_pos - (1 - tmpheight),
      lab = rev(tick_pos), yend = rep(.3, length(tick_pos))
    )


    p_tcr_all <- ggtree(tmp) %<+% tissue_annotations + xlim(-.45, 1.25) +
      geom_tiplab(aes(color = new_col, label = tissue_label),
        offset = 0.01, # (1-max(tmp$edge.length)) + .01,
        align = FALSE, linetype = NA
      ) +
      geom_tippoint() +
      scale_linetype(guide = "none") +
      scale_color_identity() +
      geom_segment(y = 0, yend = 0, x = min(ticks$x), xend = max(ticks$x)) +
      annotate(geom = "text", x = ticks$x, y = -1, label = ticks$lab) +
      geom_segment(data = ticks, aes(y = -yend, yend = 0, x = x, xend = x)) +
      theme(legend.position = c(.2, .95))
    mouse_summary_plist[[gr]] <- p_tcr_all
    ggsave(p_tcr_all, filename = file.path(expl_figs_dir, str_glue("aggregated_mouse_tcr_{gr}_day{cday}.pdf")), width = 6, height = 8, device = "pdf")

    tcr_agg_mice_ordered[upper.tri(tcr_agg_mice_ordered)] <- NA
    diag(tcr_agg_mice_ordered) <- NA
    p_tcr_all_heat <- t(tcr_agg_mice_ordered) %>%
      as.data.frame() %>%
      rownames_to_column("sample1") %>%
      pivot_longer(-sample1, names_to = "sample2") %>%
      rename(thisfill = value) %>%
      filter(!is.na(thisfill)) %>%
      filter(sample1 != sample2) %>%
      mutate(
        #       sample1=factor(sample1, levels = samples_ordered_clean, ordered = TRUE),
        #        sample2=rev(factor(sample2, levels = samples_ordered_clean, ordered = TRUE)),
        lab = round(thisfill, 2)
      ) %>%
      ggplot(aes(x = sample1, y = sample2, fill = thisfill)) +
      geom_tile() +
      coord_fixed(ratio = 1, ) +
      geom_text(aes(label = lab), color = "grey50", size = 1) +
      scale_fill_viridis_c(option = "A") +
      guides(fill = guide_colourbar(direction = "horizontal")) +
      scale_y_discrete(position = "left", expand = c(0, 0), limits = rev(samples_ordered_clean)) +
      #      scale_y_discrete(position = "right", expand = c(0,0), breaks=samples_ordered_clean) +
      scale_x_discrete(expand = c(0, 0), limits = samples_ordered_clean) +
      labs(x = "", y = "", fill = "Mean JSD") +
      theme(
        axis.text.x = element_text(angle = 45, hjust = 1, size = 9),
        legend.position = c(.7, .7),
      )

    ggsave(p_tcr_all_heat,
      filename = file.path(expl_figs_dir, str_glue("aggregated_mouse_tcr_{gr}_day{cday}_heatmap.pdf")), width = 6, height = 8, device = "pdf"
    )
    ggsave(p_tcr_all_heat + scale_fill_gradient2(midpoint = .5, limits = c(0, 1)),
      filename = file.path(expl_figs_dir, str_glue("aggregated_mouse_tcr_{gr}_day{cday}_heatmap_altcolor.pdf")), width = 6, height = 8, device = "pdf"
    )
  }
}
dir.create("figures/fig6/", showWarnings = FALSE)
file.copy(
  file.path(expl_figs_dir, str_glue("aggregated_mouse_tcr_GvHD_day7_heatmap.pdf")),
  file.path("figures/fig6/", str_glue("6A1-aggregated_mouse_tcr_GvHD_day7_heatmap.pdf"))
)
file.copy(
  file.path(expl_figs_dir, str_glue("aggregated_mouse_tcr_GvHD_day14_heatmap.pdf")),
  file.path("figures/fig6/", str_glue("6A2-aggregated_mouse_tcr_GvHD_day14_heatmap.pdf"))
)
```


## Cumulative abundancy by harvest day

```{r}
top_mouse_data <- tcr_raw %>%
  filter(org == "Mouse") %>%
  filter(grepl("Balb|Recipient", sample_name)) %>%
  mutate(harvest_day = ifelse(is.na(collection_day), 0, collection_day)) %>%
  group_by(sample_name) %>%
  mutate(rank = rank(desc(productive_frequency), ties.method = "min")) %>%
  ungroup() %>%
  filter(rank <= 10) %>%
  mutate(tissue = ifelse(tissue == "terminalileum", "ileum", tissue)) %>%
  group_by(sample_name, indv_label, tissue, harvest_day) %>%
  summarize(sum_prod_freq = sum(productive_frequency))

p_mouse_cumutop10 <- top_mouse_data %>%
  ggplot(aes(x = harvest_day, fill = as.character(harvest_day), y = sum_prod_freq, group = harvest_day)) +
  geom_boxplot() +
  # geom_jitter(height = 0, alpha = .4, size = 2) +
  scale_x_continuous(breaks = c(0, 7, 14), labels = c("control", "7", "14")) +
  scale_fill_manual(values = c("control" = "#CCD2D0", "14" = "#B6DFE9", "7" = "#172470")) +
  facet_grid(~tissue) +
  labs(x = "", y = "sum productive frequency \n of top 10 clones", fill = "Harvest Day") +
  theme_bw()+
    theme(axis.text.x = element_blank())


ggsave(p_mouse_cumutop10, filename = "figures/abundance/S8F-mouse-top-abundances.pdf", width = 8, height = 5)

top_mouse_data
mod.slope1 <- lmerTest::lmer(sum_prod_freq ~
harvest_day +
  tissue +
  (1 | indv_label),
data = top_mouse_data %>% filter(harvest_day > 0) %>% mutate(harvest_day = as.character(harvest_day))
)
summary(mod.slope1)
```
